+++ Source/SwiftLintBuiltInRules/Models/BuiltInRules.swift
@@ -126,7 +126,7 @@ public let builtInRules: [any Rule.Type] = [
     NSObjectPreferIsEqualRule.self,
     NestingRule.self,
     NimbleOperatorRule.self,
-    NoEmptyFunctionRule.self,
+    NoEmptyFunctionBodyRule.self,
     NoExtensionAccessModifierRule.self,
     NoFallthroughOnlyRule.self,
     NoGroupingExtensionRule.self,