+++ Source/SwiftLintBuiltInRules/Rules/Style/VerticalWhitespaceOpeningBracesRule.swift
@@ -141,8 +141,6 @@ struct VerticalWhitespaceOpeningBracesRule: Rule {
 }
 
 extension VerticalWhitespaceOpeningBracesRule: OptInRule {
-    init(configuration: Any) throws { /* No Body */ }
-
     static let description = RuleDescription(
         identifier: "vertical_whitespace_opening_braces",
         name: "Vertical Whitespace after Opening Braces",