+++ CHANGELOG.md
@@ -24,9 +24,7 @@
 * Linting got around 20% faster due to the praisworthy performance
   improvements done in the [SwiftSyntax](https://github.com/apple/swift-syntax)
   library.
-*  Add `non_empty_function` opt-in rule to validate that functions should not have empty body. Otherwise show warning to write comments.
-*  Add `non_empty_function` opt-in rule to validate that `function`, `init` and `deinit` should not have empty body. Otherwise show warning to ask to write comments.
-*  Add `non_empty_function` default rule to validate that `function`, `init` and `deinit` should not have empty body. Otherwise show warning to ask to write comments.
+*  Add `non_empty_function_body` default rule to validate that `function`, `init` and `deinit` should not have empty body. Otherwise show warning to ask to write comments.
     [Ueeek](https://github.com/Ueeek)
     [#5615](https://github.com/realm/SwiftLint/issues/5615)